Added the ability to read .ini input files & create new directories.
This update has two significant additions to the code: (1) In addition to the usual config file, one can now equally supply the input parameters through a .ini file. Internally, the .ini file will be converted into a config file (with .cfg extension, to distinguish it from the original .config) and the code will continue to work with the config format internally. (2) If the user had defined any directories that do not already exist, the program now automatically creates these directories instead of crashing.